Revert "FLINK-30893 Replace not specifying a driver in Minikube by something"
This reverts commit 8c8554b6

+++ flink-end-to-end-tests/test-scripts/common_kubernetes.sh
@@ -24,7 +24,7 @@ CONTAINER_SCRIPTS=${END_TO_END_DIR}/test-scripts/container-scripts
 MINIKUBE_START_RETRIES=3
 MINIKUBE_START_BACKOFF=5
 RESULT_HASH="e682ec6622b5e83f2eb614617d5ab2cf"
-MINIKUBE_VERSION="latest"
+MINIKUBE_VERSION="v1.28.0"
 
 NON_LINUX_ENV_NOTE="****** Please start/stop minikube manually in non-linux environment. ******"
 
@@ -43,7 +43,7 @@ function setup_kubernetes_for_linux {
             chmod +x kubectl && sudo mv kubectl /usr/local/bin/
     fi
     # Download minikube when it is not installed beforehand.
-    if [ -x "$(command -v minikube)" ]; then
+    if [ -x "$(command -v minikube)" ] && [[ "$(minikube version | grep -c $MINIKUBE_VERSION)" == "0" ]]; then
       echo "Removing any already installed minikube binaries ..."
       sudo rm "$(which minikube)"
     fi
@@ -51,6 +51,13 @@ function setup_kubernetes_for_linux {
     if ! [ -x "$(command -v minikube)" ]; then
       echo "Installing minikube $MINIKUBE_VERSION ..."
       curl -Lo minikube https://storage.googleapis.com/minikube/releases/$MINIKUBE_VERSION/minikube-linux-$arch && \
+          chmod +x minikube && sudo mv minikube /usr/bin/minikube
+    fi
+
+    # Install the latest version, avoid none driver of minikube
+    if ! [ -x "$(command -v minikube)" ]; then
+      echo "Installing latest version minikube ..."
+      curl -Lo minikube https://storage.googleapis.com/minikube/releases/latest/minikube-linux-$arch && \
           chmod +x minikube && sudo mv minikube /usr/bin/minikube
     fi
